+++ arm/phfs.c
@@ -1,4 +1,4 @@
-/*
+/* 
  * Phoenix-RTOS
  *
  * plo - operating system loader
@@ -37,7 +37,7 @@
 
 struct {
 	s32 (*open)(u16, char *, u32);
-	s32 (*read)(u16, s32, u32 *, u8 *, u32);
+	s32 (*read)(u16, s32, u64 *, u8 *, u32);
 	s32 (*close)(u16, s32);
 	unsigned int dn;
 } phfs_handlers[2];
@@ -47,12 +47,12 @@ s32 phfs_open(u16 pdn, char *name, u32 flags)
 {
 	if (pdn > 2)
 		return ERR_ARG;
-
+	
 	return phfs_handlers[pdn].open(phfs_handlers[pdn].dn, name, flags);
 }
 
 
-s32 phfs_read(u16 pdn, s32 handle, u32 *pos, u8 *buff, u32 len)
+s32 phfs_read(u16 pdn, s32 handle, u64 *pos, u8 *buff, u32 len)
 {
 	if (pdn > 2)
 		return ERR_ARG;
@@ -65,7 +65,7 @@ s32 phfs_close(u16 pdn, s32 handle)
 {
 	if (pdn > 2)
 		return ERR_ARG;
-
+	
 	return phfs_handlers[pdn].close(phfs_handlers[pdn].dn, handle);
 }
 
@@ -73,7 +73,7 @@ s32 phfs_close(u16 pdn, s32 handle)
 void phfs_init(void)
 {
 	unsigned int i;
-
+	
 	/* Handlers for com devices */
 	phfs_handlers[PDN_FLASH].open = flash_open;
 	phfs_handlers[PDN_FLASH].read = flash_read;
@@ -83,6 +83,6 @@ void phfs_init(void)
 	phfs_handlers[PDN_MMCBLK].read = mmcblk_read;
 	phfs_handlers[PDN_MMCBLK].close = mmcblk_close;
 	phfs_handlers[PDN_MMCBLK].dn = 0;
-
+	
 	return;
 }
